Working values and stack initialization

+++ example.c
@@ -1,12 +1,14 @@
+#include <stdint-gcc.h>
 #include "stdio.h"
 #include "polynomial_degree.c"
+#include "limits.h"
 
 int main() {
     const int* str;
     size_t n = 100;
-    int res = polynomial_degree(str, n);
+    uint64_t res = polynomial_degree(str, n);
 
 
-    printf("res = %d\n", res);
+    printf("res = %lu\n", res);
 
 }
